% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/confmatrix.R
\name{confmatrix}
\alias{confmatrix}
\title{A confusion Matrix Function}
\usage{
confmatrix(table, revPositive = FALSE)
}
\arguments{
\item{table}{A confusion matrix table}
}
\description{
This function allows you to calculate prediction performance measures from a confusion matrix.
}
\examples{
matrix = rbind(c(30,60),c(10,90))
confmatrix(matrix)
}
\keyword{confustion}
\keyword{matrix}

